+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Features:
 - Frontend timespec completely reworked.
 - Frontend evaluation logic moved into separate component.
 - Created file format 2.0, which is better at not reading data that is not requested during expression evaluation.
+- A new [tcp collector](doc/collectors/tcp.md) can be used to monitor if a port is accepting connections.
 
 Bug fixes:
 - Fix RPC timeout caused by BufferedIterator reading underlying iterator in the constructor.  This caused RPC calls to time out if the underlying iterator required more than 30 seconds for its first element to be available.